add read.py
