+++ tasks/requests/core/__init__.py
@@ -1,0 +1,2 @@
+from .database import WikiDatabase,Query
+from .module import RequestsPage,RequestsScanner,PageProcessor
